add expert requests

+++ db/migrate/20200913101359_create_parameter_methodology_total_scores.rb
@@ -3,7 +3,7 @@ class CreateParameterMethodologyTotalScores < ActiveRecord::Migration[6.0]
     create_table :parameter_methodology_total_scores do |t|
       t.belongs_to :parameter_value
       t.belongs_to :methodology
-      t.integer :score
+      t.decimal :score
     end
   end
 end
